Graphics.Rendering.Chart.Gtk

Description

Documentation

renderableToWindow :: Renderable a -> Int -> Int -> IO ()

Display a renderable in a gtk window.

Note that this is a convenience function that initialises GTK on its first call, but not subsequent calls. Hence it's unlikely to be compatible with other code using gtk. In that case use createRenderableWindow.

toWindow :: (Default r, ToRenderable r) => Int -> Int -> EC r () -> IO ()

Generate a new GTK window from the state content of an EC computation. The state may have any type that is an instance of ToRenderable

createRenderableWindow :: Renderable a -> Int -> Int -> IO Window

Create a new GTK window displaying a renderable.

updateCanvas :: Renderable a -> DrawingArea -> IO Bool